pyproject.toml
[project]
name = "zenn-hackathon-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.10"
dependencies = [
    "google-cloud-texttospeech>=2.24.0",
]
